Light scatter: Step 3 - Refine sample locations
Hope it works, had to do some ugly quick fixes to get it up and running.
Also it's quite hard (read impossible) to verify if shader works
correctly because apparently you can't render the texture because it has
R16G16 format? Somewhat tricky to understand even though the compute
shader syntax is easy. If something doesn't work out we can try to not
use refinement at first because it might not be crucial.