More useful OBJ-loader + tiny bug :)
OBJ loader now handles batches better. Didn't bother to fix material
because it's not used anyway. Should also work with negative face
indices. It turns out the test model I was using was messed up so I
can't know for sure. But the main thing I wanted is done and it should
be a good foundation if further work is required. Oh, and I also found a
tiny bug in the directional light shader that was fixed by a mighty
minus.