TiledDeferred: A first step. Read description.
First of all: performance drop. This has to do with two things. The
first is that the old lights are still rendered, but that will of course
be changed when tiled is up and running (duh). The second reason is that
culling is not yet implemented, which means that all lights are
calculated for every pixel.
On another note: for now, the system has some hard coded values and only
uses point lights.